+++ 3_3.py
@@ -1,14 +1,14 @@
 name = input("Please, input your name: ")
 while True:
     try:
-        age = int(input("Please, input your age: "))
+        age = float(input("Please, input your age: "))
     except ValueError:
         print("it's not a number.")
     else:
         break
-age = str(age)
+age = str(int(age))
 city = input("Please, input your city: ")
-# opt1
+# opt 1
 print('Hello ' + name + ' ' + age + ' y.o. from ' + city)
 # opt 2
 print('Hi, are you %s %s y.o. from %s ?' % (name, age, city))
